Backend/requirement.txt
fastapi
uvicorn
numpy
tensorflow
pillow
python-multipart